Working on getting buttons to function.

+++ DirectX11_Starter_2015/DirectX11_Starter/Input.h
@@ -49,8 +49,8 @@ public:
 	static bool GetMouseRightBtnDownThisFrame() { return GetInstance()->mouseState.rightDown && !GetInstance()->mouseState.rightPrevFrameDown; }
 	static int GetMousePositionX() { return GetInstance()->mouseState.screenX; }
 	static int GetMousePositionY() { return GetInstance()->mouseState.screenY; }
-	static float GetMousePositionNormalizedX() { return GetInstance()->mouseState.screenX / GetInstance()->mouseState.screenWidth; }
-	static float GetMousePositionNormalizedY() { return GetInstance()->mouseState.screenY / GetInstance()->mouseState.screenHeight; }
+	static float GetMousePositionNormalizedX() { return GetInstance()->mouseState.screenX / (float)GetInstance()->mouseState.screenWidth; }
+	static float GetMousePositionNormalizedY() { return GetInstance()->mouseState.screenY / (float)GetInstance()->mouseState.screenHeight; }
 	static int GetMouseDeltaX() { return GetInstance()->mouseState.screenX - GetInstance()->mouseState.previousFrameScreenX; }
 	static int GetMouseDeltaY() { return GetInstance()->mouseState.screenY - GetInstance()->mouseState.previousFrameScreenY; }
 	//KeyboardStuff
